apd components refactor
